chore(deps): fix audit vulns and migrate routes to Next 15 async params
- Bump turbo 2.9.9 → 2.9.14 (login CSRF, Yarn Berry RCE)
- Override transitive deps: ws ≥8.20.1, brace-expansion ≥5.0.6,
  postcss ≥8.5.10, @xmldom/xmldom ≥0.8.13, @tootallnate/once ≥3.0.1
- Migrate dynamic route handlers to Next 15 signature: params is now
  Promise<...>, awaited at handler entry
- Update createApiRoute wrapper to match Next 15 RouteContext
- Extract cache-stats buildResponse to sibling aggregate.ts since
  Next 15 forbids non-handler exports from route files

bun audit: 0 vulns (was 11). typecheck/lint/test all clean.
