+++ apps/web/src/app/api/calendar/sources/[sourceId]/route.ts
@@ -5,10 +5,10 @@ import {
 
 export const dynamic = "force-dynamic";
 
-export async function PATCH(request: Request, context: { params: { sourceId: string } }) {
+export async function PATCH(request: Request, context: { params: Promise<{ sourceId: string }> }) {
   return schedulesPatch(request, context);
 }
 
-export async function DELETE(request: Request, context: { params: { sourceId: string } }) {
+export async function DELETE(request: Request, context: { params: Promise<{ sourceId: string }> }) {
   return schedulesDelete(request, context);
 }
