+++ apps/web/src/app/api/feed/[postId]/comments/route.ts
@@ -5,7 +5,7 @@ import { validateJson, validationErrorResponse, ValidationError } from "@/lib/se
 import { checkRateLimit, buildRateLimitResponse } from "@/lib/security/rate-limit";
 import { getOrgMembership } from "@/lib/auth/api-helpers";
 
-export async function GET(request: NextRequest, { params }: { params: { postId: string } }) {
+export async function GET(request: NextRequest, { params }: { params: Promise<{ postId: string }> }) {
   try {
     const supabase = await createClient();
     const {
@@ -26,7 +26,7 @@ export async function GET(request: NextRequest, { params }: { params: { postId: 
       return buildRateLimitResponse(rateLimit);
     }
 
-    const { postId } = params;
+    const { postId } = await params;
 
     // Fetch post to verify it exists and get org_id
     const { data: post } = await supabase
@@ -69,7 +69,7 @@ export async function GET(request: NextRequest, { params }: { params: { postId: 
   }
 }
 
-export async function POST(request: NextRequest, { params }: { params: { postId: string } }) {
+export async function POST(request: NextRequest, { params }: { params: Promise<{ postId: string }> }) {
   try {
     const supabase = await createClient();
     const {
@@ -91,7 +91,7 @@ export async function POST(request: NextRequest, { params }: { params: { postId:
       return buildRateLimitResponse(rateLimit);
     }
 
-    const { postId } = params;
+    const { postId } = await params;
     const { body } = await validateJson(request, createCommentSchema);
 
     // Fetch post to check it exists and get org_id
