+++ package.json
@@ -35,6 +35,11 @@
   "overrides": {
     "glob": ">=10.4.6",
     "@types/react": "^18",
-    "@types/react-dom": "^18"
+    "@types/react-dom": "^18",
+    "ws": ">=8.20.1",
+    "brace-expansion": ">=5.0.6",
+    "postcss": ">=8.5.10",
+    "@xmldom/xmldom": ">=0.8.13",
+    "@tootallnate/once": ">=3.0.1"
   }
 }
